refactor: centralize command handlers map
