+++ src/commandsHandler/menuHandler.js
@@ -5,54 +5,16 @@ const {
   MENU_CALLBACK_TYPE,
   COMMAND_BEST_TEAMS,
   COMMAND_CURRENT_TEAM_INFO,
-  COMMAND_CHIPS,
   COMMAND_PRINT_CACHE,
   COMMAND_RESET_CACHE,
   COMMAND_HELP,
-  COMMAND_TRIGGER_SCRAPING,
-  COMMAND_LOAD_SIMULATION,
-  COMMAND_GET_CURRENT_SIMULATION,
-  COMMAND_GET_BOTFATHER_COMMANDS,
   COMMAND_NEXT_RACE_INFO,
-  COMMAND_BILLING_STATS,
-  COMMAND_VERSION,
-  COMMAND_SET_LANGUAGE,
 } = require('../constants');
 
-// Import command handlers directly to avoid circular dependency
-const { handleBestTeamsMessage } = require('./bestTeamsHandler');
-const { handleChipsMessage } = require('./chipsHandler');
-const { calcCurrentTeamInfo } = require('./currentTeamInfoHandler');
-const { handleGetBotfatherCommands } = require('./getBotfatherCommandsHandler');
-const { handleGetCurrentSimulation } = require('./getCurrentSimulationHandler');
+// Import command handlers directly to avoid circular dependency in mapping
 const { displayHelpMessage } = require('./helpHandler');
-const { handleLoadSimulation } = require('./loadSimulationHandler');
-const { handleNextRaceInfoCommand } = require('./nextRaceInfoHandler');
-const { sendPrintableCache } = require('./printCacheHandler');
-const { resetCacheForChat } = require('./resetCacheHandler');
-const { handleScrapingTrigger } = require('./scrapingTriggerHandler');
-const { handleBillingStats } = require('./billingStatsHandler');
-const { handleVersionCommand } = require('./versionHandler');
-const { handleSetLanguage } = require('./setLanguageHandler');
+const { COMMAND_HANDLERS } = require('./commandHandlers');
 const { t } = require('../i18n');
-
-// Map commands to their handler functions
-const COMMAND_HANDLERS = {
-  [COMMAND_BEST_TEAMS]: handleBestTeamsMessage,
-  [COMMAND_CURRENT_TEAM_INFO]: calcCurrentTeamInfo,
-  [COMMAND_CHIPS]: handleChipsMessage,
-  [COMMAND_PRINT_CACHE]: sendPrintableCache,
-  [COMMAND_RESET_CACHE]: resetCacheForChat,
-  [COMMAND_HELP]: displayHelpMessage,
-  [COMMAND_TRIGGER_SCRAPING]: handleScrapingTrigger,
-  [COMMAND_LOAD_SIMULATION]: handleLoadSimulation,
-  [COMMAND_GET_CURRENT_SIMULATION]: handleGetCurrentSimulation,
-  [COMMAND_GET_BOTFATHER_COMMANDS]: handleGetBotfatherCommands,
-  [COMMAND_NEXT_RACE_INFO]: handleNextRaceInfoCommand,
-  [COMMAND_BILLING_STATS]: handleBillingStats,
-  [COMMAND_VERSION]: handleVersionCommand,
-  [COMMAND_SET_LANGUAGE]: handleSetLanguage,
-};
 
 async function displayMenuMessage(bot, msg) {
   const chatId = msg.chat.id;
